[스탯 변경 사항 배틀에 적용 및 UI 표기 오류 해결]

+++ Assets/97. Moon/V1/Scripts/Stat.cs
@@ -25,7 +25,7 @@ public class Stat
     [Header("Calculated Results")]
     public int max_Hp;
     public int current_Hp;
-    public int def;
+    public float def;
     public int atk;
 
     public int speed;
